Parse JSON numbers to Number type instead of Double.
Currently all numbers are parsed as Double thus integers are parsed as
"int.0", which is **wrong** according to JSON specification (see
www.json.org). With the spec, JSON's number is defined as "int or
int.frac or scientific annotation". This fix parses JSON numbers to
Java's Number using NumberFormat parser, which correctly parses JSON
numbers to Long and Double respectively. This resolves a lot of issues
raised in the past years and also correctly parse "9223372036854775807"
to Long.MAX_VALUE and no precision is lost.

Fuck gson developers (#1084 and many others)